Fix compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@24305 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/LogDialog/LogDlg.cpp
@@ -6624,7 +6624,7 @@ bool CLogDlg::CheckMultipleDiffs( UINT selCount )
 void CLogDlg::ExecuteMultipleDiffChangedPaths(ContextMenuInfoForChangedPathsPtr pCmi)
 {
     INT_PTR selIndex = 0;
-    int nPaths = pCmi->ChangedLogPathIndices.size();
+    int nPaths = (int)pCmi->ChangedLogPathIndices.size();
    
     // warn if we exceed Software\\TortoiseSVN\\NumDiffWarning or 15 if not set
     if (!CheckMultipleDiffs(nPaths))
